Tweet compose complete

+++ twitterClone/Models/TwitterUser.swift
@@ -13,8 +13,8 @@ struct TweetUser: Codable {
     let id: String
     var displayName: String = ""
     var username: String = ""
-    var followersCount: Double = 0
-    var followingCount: Double = 0
+    var followersCount: Int = 0
+    var followingCount: Int = 0
     var createdON: Date = Date()
     var bio: String = ""
     var avatarPath: String = ""
